+++ R/check_data.R
@@ -11,7 +11,9 @@
 #'  * `PLINK` text data: use \link{check_plink_text_data}
 #'  * `PLINK` binary data: use \link{check_plink_bin_data}
 #'  * `PLINK2` binary data: use \link{check_plink2_bin_data}
-#'  * any of these: use \link{check_data}
+#'  * any `PLINK`/`PLINK2` data: use \link{check_data}
+#'  * quantitative trait analysis data: use \link{check_assoc_qt_data}
+#'
 #' @examples
 #' check_data(data = create_test_plink_text_data())
 #' check_data(data = create_test_plink_bin_data())
@@ -52,6 +54,9 @@ check_data <- function(data) {
     return(invisible(data))
   }
   if (plinkr::is_plink2_bin_filenames(plink2_bin_filenames = data)) {
+    return(invisible(data))
+  }
+  if (plinkr::is_assoc_qt_data(assoc_qt_data = data)) {
     return(invisible(data))
   }
   stop(