+++ R/is_assoc_qt_data.R
@@ -1,8 +1,11 @@
-#' Determine if the `plink_text_data` is indeed `PLINK` text data
+#' Determine if the `assoc_qt_data` is indeed data used for a
+#' quantitative trait analysis (as done by \link{assoc_qt})
 #'
-#' Determine if the `plink_text_data` is indeed `PLINK` text data
+#' Determine if the `assoc_qt_data` is indeed data used for a
+#' quantitative trait analysis (as done by \link{assoc_qt})
 #' @inheritParams default_params_doc
-#' @return \link{TRUE} if the `plink_text_data` is indeed `PLINK` text data
+#' @return \link{TRUE} if the `assoc_qt_data` is indeed data used for a
+#' quantitative trait analysis (as done by \link{assoc_qt})
 #' @seealso `plinkr` has multiple functions to test if a type of data is valid:
 #'
 #'  * `PLINK` text data: use \link{is_plink_text_data}
@@ -12,19 +15,19 @@
 #'  * associative trait analysis data: use \link{is_assoc_qt_data}
 #'
 #' @examples
-#' plink_text_data <- create_plink_text_data(
+#' assoc_qt_data <- create_assoc_qt_data(
 #'   ped_table = get_test_ped_table(),
 #'   map_table = get_test_map_table()
 #' )
-#' is_plink_text_data(plink_text_data)
+#' is_assoc_qt_data(assoc_qt_data)
 #' @author Richèl J.C. Bilderbeek
 #' @export
-is_plink_text_data <- function(
-  plink_text_data
+is_assoc_qt_data <- function(
+  assoc_qt_data
 ) {
   result <- FALSE
   tryCatch({
-    plinkr::check_plink_text_data(plink_text_data = plink_text_data)
+    plinkr::check_assoc_qt_data(assoc_qt_data = assoc_qt_data)
     result <- TRUE
   }, error = function(e) {} # nolint indeed ignore error
   )