+++ R/is_plink_bin_data.R
@@ -6,6 +6,8 @@
 #'  * `PLINK` text data: use \link{is_plink_text_data}
 #'  * `PLINK` binary data: use \link{is_plink_bin_data}
 #'  * `PLINK2` binary data: use \link{is_plink2_bin_data}
+#'  * any type of `PLINK`/`PLINK2` data: use \link{is_data}
+#'  * associative trait analysis data: use \link{is_assoc_qt_data}
 #'
 #' @examples
 #' plink_bin_data <- create_plink_bin_data(